pyproject.toml
[tool.poetry]
name = "airbnb-backend"
version = "0.1.0"
description = ""
authors = ["acecic82 <102938al@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
